--- my-app/react-app/src/components/models/Product.ts
@@ -1,9 +1,0 @@
-export interface Product {
-  id: number
-  name: string
-  price: number
-  description: string
-  category: string
-  quantity?: number
-  imageURL?: string
-}
