+++ my-app/react-app/src/containers/SignInOutContainer.tsx
@@ -1,8 +1,8 @@
 import { Paper, Tabs, Box, Tab } from '@mui/material'
 import React from 'react'
-import Login from '../components/login/Components/login'
-import Signup from '../components/login/Components/signup'
-import TabPanel from '../components/misc/TabPanel'
+import Login from '../components/login/login'
+import Signup from '../components/login/signup'
+import TabPanel from '../misc/TabPanel'
 
 const SignInOutContainer: React.FC = () => {
   const [value, setValue] = React.useState(0)
