+++ my-app/react-app/src/models/LoggingModel.ts
