+++ my-app/react-app/src/models/Product.ts
